entity-resolution/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Record linkage
recordlinkage>=0.15

# Machine learning
scikit-learn>=1.3.0

# String similarity
jellyfish>=1.0.0
python-Levenshtein>=0.21.0

# Utilities
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
